Pie chart representation/requirements.txt
azure-core==1.21.1
azure-storage-blob==12.9.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
Flask==2.0.2
idna==3.3
isodate==0.6.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
msrest==0.6.21
numpy==1.22.1
oauthlib==3.2.0
pandas==1.4.0
pycparser==2.21
pymssql==2.2.4
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
urllib3==1.26.8
Werkzeug==2.0.2
